feat: Cell Text 변경로직 DTO로 이관
